Add quickstart section in readme
