+++ src/python/ap_features/testing.py
@@ -1,0 +1,19 @@
+import numpy as np
+
+
+def ca_transient(t, tstart=0.05):
+    tau1 = 0.05
+    tau2 = 0.110
+
+    ca_diast = 0.0
+    ca_ampl = 1.0
+
+    beta = (tau1 / tau2) ** (-1 / (tau1 / tau2 - 1)) - (tau1 / tau2) ** (-1 / (1 - tau2 / tau1))
+    ca = np.zeros_like(t)
+
+    ca[t <= tstart] = ca_diast
+
+    ca[t > tstart] = (ca_ampl - ca_diast) / beta * (
+        np.exp(-(t[t > tstart] - tstart) / tau1) - np.exp(-(t[t > tstart] - tstart) / tau2)
+    ) + ca_diast
+    return ca
