+++ tests/test_features.py
@@ -129,20 +129,6 @@ def test_all_cost_terms(synthetic_data):
     assert np.all(cost - expected_cost < 1e-10)
 
 
-# def test_cost_terms_trace(synthetic_data):
-#     arr, t, expected_cost = synthetic_data
-#     V = np.ascontiguousarray(arr[0, :])
-
-#     cost_terms_c = apf.cost_terms_trace(V, t, backend="c")
-#     cost_terms_py = apf.cost_terms_trace(V, t, backend="numba")
-
-#     lst = apf.list_cost_function_terms_trace()
-#     inds = np.where(["int_30" in item for item in lst])[0]
-#     x = np.delete(np.arange(len(lst)), inds)
-
-#     assert np.all(np.abs(cost_terms_c[x] - cost_terms_py[x]) < 1e-10)
-
-
 @pytest.mark.parametrize("factor, backend", it.product((40, 60, 80), ("numba",)))
 def test_apd_equivalence(factor, backend, synthetic_data):
     arr, t, expected_cost = synthetic_data
@@ -152,16 +138,6 @@ def test_apd_equivalence(factor, backend, synthetic_data):
 
     # We expect some difference here, but no more than 1ms
     assert abs(apd_x - apd_py) < 1
-
-
-# @pytest.mark.parametrize("factor", (40, 60, 80))
-# def test_apd_equivalence_c_numba(factor, synthetic_data):
-#     arr, t, expected_cost = synthetic_data
-#     V = np.ascontiguousarray(arr[0, :])
-#     apd_numba = apf.apd(V=V, t=t, factor=factor, backend="numba")
-#     apd_c = apf.apd(V=V, t=t, factor=factor, backend="c")
-
-#     assert abs(apd_c - apd_numba) < 1e-10
 
 
 @pytest.mark.parametrize(
